% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apprisHuman}
\alias{apprisHuman}
\title{APPRIS annotation for human genes (Ensembl release 104)}
\format{
data.frame
}
\usage{
apprisHuman
}
\description{
data.frame containing APPRIS annotation for human genes
using the APPRIS release 47 and the Ensembl release 104
gene annotation.
}
\keyword{datasets}
