% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apprisMouse}
\alias{apprisMouse}
\title{APPRIS annotation for mouse genes (Ensembl release 102)}
\format{
data.frame
}
\usage{
data(apprisMouse, package="crisprDesignData")
}
\description{
data.frame containing APPRIS annotation for mouse genes
using the APPRIS release 47 and the Ensembl release 102
gene annotation.
}
\keyword{datasets}
