% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canonicalCyno}
\alias{canonicalCyno}
\title{Canonical Ensembl transcripts for crab-eating macaque genes (Ensembl release 108)}
\format{
data.frame
}
\usage{
data(canonicalCyno, package="crisprDesignData")
}
\description{
data.frame containing canonical Ensembl transcripts for crab-eating macaque 
genes. First column contains transcript ID, second column contains gene ID.
}
\keyword{datasets}
