% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canonicalHuman}
\alias{canonicalHuman}
\title{Canonical Ensembl transcripts for human genes (Ensembl release 104)}
\format{
data.frame
}
\usage{
canonicalHuman
}
\description{
data.frame containing canonical Ensembl transcripts for human genes.
First column contains transcript ID, second column contains gene ID.
}
\keyword{datasets}
