% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canonicalMouse}
\alias{canonicalMouse}
\title{Canonical Ensembl transcripts for mouse genes (Ensembl release 102)}
\format{
data.frame
}
\usage{
data(canonicalMouse, package="crisprDesignData")
}
\description{
data.frame containing canonical Ensembl transcripts for mouse genes.
First column contains transcript ID, second column contains gene ID.
}
\keyword{datasets}
