% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr.repeats.mm10}
\alias{gr.repeats.mm10}
\title{Repeat elements coordinates for mm10}
\format{
GRanges
}
\usage{
gr.repeats.mm10
}
\description{
Repeat elements coordinates for mm10
}
\keyword{datasets}
