% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanCagePeaks}
\alias{humanCagePeaks}
\title{Genomic coordinates of FANTOM5 CAGE Peaks in hg38 coordinates}
\format{
GRanges
}
\usage{
humanCagePeaks
}
\description{
Genomic coordinates of FANTOM5 CAGE Peaks in hg38 coordinates.
}
\keyword{datasets}
