% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrnasHuman}
\alias{mrnasHuman}
\title{DNAStringSet of mRNA sequences of human transcripts (Ensembl)}
\format{
DNAStringSet
}
\usage{
mrnasHuman
}
\description{
DNAStringSet of mRNA sequences of human transcripts
(Ensembl release 104) extracted from the txdb_human object.
}
\keyword{datasets}
