% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mrnasMouse}
\alias{mrnasMouse}
\title{DNAStringSet of mRNA sequences of mouse transcripts (Ensembl)}
\format{
DNAStringSet
}
\usage{
mrnasMouse
}
\description{
DNAStringSet of mRNA sequences of mouse transcripts
(Ensembl release 102) extracted from the txdb_human object.
}
\keyword{datasets}
