% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfamTableHuman}
\alias{pfamTableHuman}
\title{Table of Pfam doains for human transcripts}
\format{
DataFrame
}
\usage{
data(pfamTableHuman, package="crisprDesignData")
}
\description{
DataFrame containing Pfam domain information for human Ensembl transcripts.
(Ensembl release 104)).
}
\keyword{datasets}
