% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pfamTableMouse}
\alias{pfamTableMouse}
\title{Table of Pfam doains for mouse transcripts}
\format{
DataFrame
}
\usage{
data(pfamTableMouse, package="crisprDesignData")
}
\description{
DataFrame containing Pfam domain information for mouse Ensembl transcripts.
(Ensembl release 102)).
}
\keyword{datasets}
