% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tss_mouse}
\alias{tss_mouse}
\title{GRanges representing Ensembl-based TSS coordinates for mouse}
\format{
GRanges
}
\usage{
data(tss_mouse, package="crisprDesignData")
}
\description{
GRanges representing Ensembl-based TSS coordinates for mouse in mm10/GRCm38
coordinates (Ensembl release 102)
}
\keyword{datasets}
