% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb_cyno}
\alias{txdb_cyno}
\title{GRangesList representing Ensembl gene model coordinates for crab-eating macaque}
\format{
GRangesList
}
\usage{
data(txdb_cyno, package="crisprDesignData")
}
\description{
GRangesList representing Ensembl gene model coordinates for the crab-eating macaque
in macFas6 coordinates (Ensembl release 108)
}
\keyword{datasets}
