% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb_human}
\alias{txdb_human}
\title{GRangesList representing Ensembl gene model coordinates for human}
\format{
GRangesList
}
\usage{
txdb_human
}
\description{
GRangesList representing Ensembl gene model coordinates for human in 
hg38/GRCh38 coordinates (Ensembl release 104)
}
\keyword{datasets}
