% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{txdb_mouse}
\alias{txdb_mouse}
\title{GRangesList representing Ensembl gene model coordinates for mouse}
\format{
GRangesList
}
\usage{
data(txdb_mouse, package="crisprDesignData")
}
\description{
GRangesList representing Ensembl gene model coordinates for mouse in 
mm10/GRCm38 coordinates (Ensembl release 102)
}
\keyword{datasets}
