% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr.repeats.hg38}
\alias{gr.repeats.hg38}
\title{Repeat elements coordinates found in the human genome (hg38/GRCh38)}
\format{
GRanges
}
\usage{
gr.repeats.hg38
}
\description{
Repeat elements coordinates for hg38 obtained from the RepeatMasker track
found at the UCSC genome browser.
}
\keyword{datasets}
