% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr.repeats.mm10}
\alias{gr.repeats.mm10}
\title{Repeat elements coordinates found in the mouse genome (mm10/GRCm38)}
\format{
GRanges
}
\usage{
gr.repeats.mm10
}
\description{
Repeat elements coordinates for mm10 obtained from the RepeatMasker track
found at the UCSC genome browser.
}
\keyword{datasets}
