% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseCagePeaks}
\alias{mouseCagePeaks}
\title{Genomic coordinates of FANTOM5 CAGE Peaks in mm10 coordinates}
\format{
GRanges
}
\usage{
mouseCagePeaks
}
\description{
Genomic coordinates of FANTOM5 CAGE Peaks in mm10 coordinates.
}
\keyword{datasets}
