% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tss_human}
\alias{tss_human}
\title{GRangesList representing Ensembl gene model coordinates for human}
\format{
GRanges
}
\usage{
tss_human
}
\description{
GRanges representing Ensembl-based TSS coordinates for human in 
hg38/GRCh38 coordinates (Ensembl release 104)
}
\keyword{datasets}
